requirements.txt
aiohttp==3.9.5
asyncpg==0.29.0
playwright==1.43.0
python-dotenv==1.0.1
python-telegram-bot==21.1.1
